Re-arrange GraphQL Compiler Plugin Code
