--- graphql-compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/Constants.java
@@ -1,89 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.ballerina.stdlib.graphql.compiler;
-
-import io.ballerina.tools.diagnostics.DiagnosticSeverity;
-
-/**
- * GraphQl compiler plugin constants.
- */
-public class Constants {
-
-    // compiler plugin constants
-    public static final String PACKAGE_PREFIX = "graphql";
-    public static final String PACKAGE_ORG = "ballerina";
-
-    // resource function constants
-    public static final String RESOURCE_FUNCTION_GET = "get";
-    public static final String LISTENER_IDENTIFIER = "Listener";
-    public static final String DOUBLE_UNDERSCORES = "__";
-
-    /**
-     * Compilation errors.
-     */
-    enum CompilationError {
-        INVALID_FUNCTION("Remote methods are not allowed in GraphQL services", "GRAPHQL_101", DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE("Invalid return type for GraphQL resource function", "GRAPHQL_102",
-                            DiagnosticSeverity.ERROR),
-        INVALID_RESOURCE_INPUT_PARAM("Invalid input parameter type for GraphQL resource function", "GRAPHQL_103",
-                                     DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_NIL("A GraphQL resource function must have a return type", "GRAPHQL_104",
-                                DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_ERROR_OR_NIL("A GraphQL resource function must have a return data type", "GRAPHQL_105",
-                                         DiagnosticSeverity.ERROR),
-        INVALID_RESOURCE_FUNCTION_ACCESSOR("Only \"get\" accessor is allowed for GraphQL resource function",
-                                           "GRAPHQL_106", DiagnosticSeverity.ERROR),
-        INVALID_MULTIPLE_LISTENERS("A GraphQL service cannot be attached to multiple listeners", "GRAPHQL_107",
-                                   DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_ERROR("A GraphQL resource function must have a return data type", "GRAPHQL_108",
-                                  DiagnosticSeverity.ERROR),
-        INVALID_LISTENER_INIT("http:Listener and graphql:ListenerConfiguration are mutually exclusive", "GRAPHQL_109",
-                              DiagnosticSeverity.ERROR),
-        INVALID_RETURN_TYPE_MULTIPLE_SERVICES("GraphQL union type cannot consist non-distinct services", "GRAPHQL_110",
-                                              DiagnosticSeverity.ERROR),
-        INVALID_FIELD_NAME(
-                "A GraphQL field name must not begin with \"__\", which is reserved by GraphQL introspection",
-                "GRAPHQL_111", DiagnosticSeverity.ERROR);
-
-        private final String error;
-        private final String errorCode;
-        private final DiagnosticSeverity diagnosticSeverity;
-
-        CompilationError(String error, String errorCode, DiagnosticSeverity diagnosticSeverity) {
-            this.error = error;
-            this.errorCode = errorCode;
-            this.diagnosticSeverity = diagnosticSeverity;
-        }
-
-        String getError() {
-            return error;
-        }
-
-        String getErrorCode() {
-            return errorCode;
-        }
-
-        DiagnosticSeverity getDiagnosticSeverity() {
-            return this.diagnosticSeverity;
-        }
-    }
-
-    private Constants() {
-    }
-}
