+++ graphql-compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/Utils.java
@@ -21,7 +21,6 @@ package io.ballerina.stdlib.graphql.compiler;
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.api.symbols.MethodSymbol;
 import io.ballerina.compiler.api.symbols.ModuleSymbol;
-import io.ballerina.compiler.api.symbols.ObjectTypeSymbol;
 import io.ballerina.compiler.api.symbols.Qualifier;
 import io.ballerina.compiler.api.symbols.Symbol;
 import io.ballerina.compiler.api.symbols.TypeDescKind;
@@ -30,10 +29,10 @@ import io.ballerina.compiler.api.symbols.TypeSymbol;
 import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
 import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
 import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
-import io.ballerina.stdlib.graphql.compiler.Constants.CompilationError;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticFactory;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
+import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 import io.ballerina.tools.diagnostics.Location;
 
 import java.util.Optional;
@@ -42,10 +41,73 @@ import java.util.Optional;
  * Util class for the compiler plugin.
  */
 public class Utils {
+
+    private Utils() {}
+
+    // compiler plugin constants
+    public static final String PACKAGE_PREFIX = "graphql";
+    public static final String PACKAGE_ORG = "ballerina";
+
+    // resource function constants
+    public static final String RESOURCE_FUNCTION_GET = "get";
+    public static final String LISTENER_IDENTIFIER = "Listener";
+    public static final String DOUBLE_UNDERSCORES = "__";
+
+    /**
+     * Compilation errors.
+     */
+    enum CompilationError {
+        INVALID_FUNCTION("Remote methods are not allowed in GraphQL services", "GRAPHQL_101", DiagnosticSeverity.ERROR),
+        INVALID_RETURN_TYPE("Invalid return type for GraphQL resource function", "GRAPHQL_102",
+                            DiagnosticSeverity.ERROR),
+        INVALID_RESOURCE_INPUT_PARAM("Invalid input parameter type for GraphQL resource function", "GRAPHQL_103",
+                                     DiagnosticSeverity.ERROR),
+        INVALID_RETURN_TYPE_NIL("A GraphQL resource function must have a return type", "GRAPHQL_104",
+                                DiagnosticSeverity.ERROR),
+        INVALID_RETURN_TYPE_ERROR_OR_NIL("A GraphQL resource function must have a return data type", "GRAPHQL_105",
+                                         DiagnosticSeverity.ERROR),
+        INVALID_RESOURCE_FUNCTION_ACCESSOR(
+                "Only \"" + RESOURCE_FUNCTION_GET + "\" accessor is allowed for GraphQL resource function",
+                "GRAPHQL_106", DiagnosticSeverity.ERROR),
+        INVALID_MULTIPLE_LISTENERS("A GraphQL service cannot be attached to multiple listeners", "GRAPHQL_107",
+                                   DiagnosticSeverity.ERROR),
+        INVALID_RETURN_TYPE_ERROR("A GraphQL resource function must have a return data type", "GRAPHQL_108",
+                                  DiagnosticSeverity.ERROR),
+        INVALID_LISTENER_INIT("http:Listener and graphql:ListenerConfiguration are mutually exclusive", "GRAPHQL_109",
+                              DiagnosticSeverity.ERROR),
+        INVALID_RETURN_TYPE_MULTIPLE_SERVICES("GraphQL union type cannot consist non-distinct services", "GRAPHQL_110",
+                                              DiagnosticSeverity.ERROR),
+        INVALID_FIELD_NAME("A GraphQL field name must not begin with \"" + DOUBLE_UNDERSCORES +
+                                   "\", which is reserved by GraphQL introspection", "GRAPHQL_111",
+                           DiagnosticSeverity.ERROR);
+
+        private final String error;
+        private final String errorCode;
+        private final DiagnosticSeverity diagnosticSeverity;
+
+        CompilationError(String error, String errorCode, DiagnosticSeverity diagnosticSeverity) {
+            this.error = error;
+            this.errorCode = errorCode;
+            this.diagnosticSeverity = diagnosticSeverity;
+        }
+
+        String getError() {
+            return error;
+        }
+
+        String getErrorCode() {
+            return errorCode;
+        }
+
+        DiagnosticSeverity getDiagnosticSeverity() {
+            return this.diagnosticSeverity;
+        }
+    }
+
     public static boolean validateModuleId(ModuleSymbol moduleSymbol) {
         String moduleName = moduleSymbol.id().moduleName();
         String orgName = moduleSymbol.id().orgName();
-        return moduleName.equals(Constants.PACKAGE_PREFIX) && orgName.equals(Constants.PACKAGE_ORG);
+        return moduleName.equals(PACKAGE_PREFIX) && orgName.equals(PACKAGE_ORG);
     }
 
     public static MethodSymbol getMethodSymbol(SyntaxNodeAnalysisContext context,
@@ -79,9 +141,9 @@ public class Utils {
                     .filter(typeDescriptor -> typeDescriptor instanceof TypeReferenceTypeSymbol)
                     .map(typeReferenceTypeSymbol -> (TypeReferenceTypeSymbol) typeReferenceTypeSymbol)
                     .anyMatch(typeReferenceTypeSymbol ->
-                            typeReferenceTypeSymbol.getModule().isPresent()
-                                    && validateModuleId(typeReferenceTypeSymbol.getModule().get()
-                            ));
+                                      typeReferenceTypeSymbol.getModule().isPresent()
+                                              && validateModuleId(typeReferenceTypeSymbol.getModule().get()
+                                      ));
         }
 
         if (listenerType.typeKind() == TypeDescKind.TYPE_REFERENCE) {
@@ -90,15 +152,14 @@ public class Utils {
         }
 
         if (listenerType.typeKind() == TypeDescKind.OBJECT) {
-            Optional<ModuleSymbol> moduleOpt = ((ObjectTypeSymbol) listenerType).getModule();
+            Optional<ModuleSymbol> moduleOpt = listenerType.getModule();
             return moduleOpt.isPresent() && validateModuleId(moduleOpt.get());
         }
         return false;
     }
 
     public static boolean isInvalidFieldName(String fieldName) {
-        return fieldName.startsWith(Constants.DOUBLE_UNDERSCORES) ||
-                fieldName.contains("/" + Constants.DOUBLE_UNDERSCORES);
+        return fieldName.startsWith(DOUBLE_UNDERSCORES) || fieldName.contains("/" + DOUBLE_UNDERSCORES);
     }
 
     static Location getLocation(Symbol typeSymbol, Location alternateLocation) {
